requirements.txt
pandas
flask
google-cloud-storage
google-cloud-aiplatform>=1.38.0
vertexai
gunicorn
scikit-learn